+++ src/@types/validation.d.ts
@@ -1,3 +1,5 @@
+import {ValidationLevel} from "../lib";
+
 export declare interface ValidationConstrains {
 
     optional?: boolean;
@@ -35,37 +37,6 @@ export declare interface ValidationTokenFunction extends ValidationConstrains {
 export declare type ValidationToken = ValidationTokenFunction | ValidationTokenGeneric | ValidationTokenGroup;
 export declare type ValidationTokenList = Array<ValidationToken>;
 
-export declare interface ValidationRuleDescriptor {
-
-    syntax?: ValidationTokenList;
-    initial?: string;
-}
-
-export declare interface ValidationRule {
-
-    syntax: ValidationTokenList;
-    pattern: string;
-    descriptors?: {
-        [key: string]: ValidationRuleDescriptor
-    }
-}
-
-export declare interface ValidationRuleSet {
-
-    [key: string]: ValidationRule;
-}
-
-export declare interface SyntaxRuleSet {
-
-    syntax: string;
-    descriptors?: {
-        [key: string]: {
-
-            syntax?: string
-        }
-    }
-}
-
 export declare interface ValidationSyntaxNode {
     syntax: string;
     ast: ValidationToken[];
@@ -78,4 +49,11 @@ export declare interface ValidationConfiguration {
     functions: Record<string, ValidationSyntaxNode>;
     syntaxes: Record<string, ValidationSyntaxNode>;
     selectors: Record<string, ValidationSyntaxNode>;
+}
+
+export interface ValidationResult {
+
+    valid: ValidationLevel;
+    node: AstNode | Token | null;
+    error: string;
 }